requirements_core.txt
# Core dependencies for reproducing main paper results
numpy>=1.21
pandas>=1.3
scipy>=1.7
scikit-learn>=1.0
torch>=1.10
matplotlib>=3.4
pyyaml>=5.4
joblib>=1.1
tqdm>=4.62
networkx>=2.6
tabulate>=0.9.0
